fix: rename JsiSKData to JsiSkData
